Taking out the ramses handlers into a new file
