+++ tangos/input_handlers/pynbody.py
@@ -313,66 +313,6 @@ class PynbodyInputHandler(finding.PatternBasedFileDiscovery, HandlerBase):
         return estimated_part_mass
 
 
-class RamsesHOPInputHandler(PynbodyInputHandler):
-    patterns = ["output_0????"]
-    auxiliary_file_patterns = ["grp*.tag"]
-
-    def match_objects(
-        self,
-        ts1,
-        ts2,
-        halo_min,
-        halo_max,
-        dm_only=False,
-        threshold=0.005,
-        object_typetag="halo",
-        output_handler_for_ts2=None,
-    ):
-
-        f1 = self.load_timestep(ts1).dm
-        h1 = self._construct_halo_cat(ts1, object_typetag)
-
-        if output_handler_for_ts2 is None:
-            f2 = self.load_timestep(ts2).dm
-            h2 = self._construct_halo_cat(ts2, object_typetag)
-        else:
-            f2 = output_handler_for_ts2.load_timestep(ts2).dm
-            h2 = output_handler_for_ts2._construct_halo_cat(ts2, object_typetag)
-
-        bridge = pynbody.bridge.OrderBridge(f1, f2, monotonic=False)
-
-        return bridge.fuzzy_match_catalog(
-            halo_min,
-            halo_max,
-            threshold=threshold,
-            only_family=pynbody.family.dm,
-            groups_1=h1,
-            groups_2=h2,
-        )
-
-
-def RamsesAdaptaHOPInputHandler(RamsesHOPInputHandler):
-    patterns = ["output_0????"]
-    auxiliary_file_patterns = ["tree_bricks???"]
-
-    def available_object_property_names_for_timestep(
-        self, ts_extension, object_typetag
-    ):
-        f = self._load_timestep(ts_extension)
-        h = self._construct_halo_cat(ts_extension, object_typetag)
-
-        halo_attributes = list(h._halo_attributes)
-        if h._read_contamination:
-            halo_attributes.extend(h._halo_attributes_contam)
-
-        attrs = chain.from_iterable(
-            tuple(always_iterable(attr)) for (attr, _len, _dtype) in halo_attributes
-        )
-
-        # We return all properties but the ids of the particles contained in the halo
-        return [attr for attr in attrs if attr != "members"]
-
-
 class GadgetSubfindInputHandler(PynbodyInputHandler):
     patterns = ["snapshot_???"]
     auxiliary_file_patterns =["groups_???"]
@@ -631,4 +571,4 @@ class ChangaUseIDLInputHandler(ChangaInputHandler):
     halo_stat_file_class_name = "AmigaIDLStatFile"
     auxiliary_file_patterns = ["*.amiga.grp"]
 
-from . import caterpillar, eagle
+from . import caterpillar, eagle, ramsesHOP
